made changes to the api structure of the Gcircle class and made child class for each plot type that inherits specific functions from the Gcircle class.
